190720 Finish

+++ controllers/videoController.js
@@ -1,5 +1,7 @@
 import routes from "../routes";
 import Video from "../models/Video";
+import Comment from "../models/Comment";
+
 
 export const home = async (req, res) => {
     try {
@@ -63,14 +65,15 @@ export const postUpload = async (req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator:req.user.id
     });
-    console.log(newVideo);
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id));
 };
 
 export const videoDetail = async (req, res) => {
-
     const {
         params: {
             id
@@ -78,7 +81,9 @@ export const videoDetail = async (req, res) => {
     } = req;
 
     try {
-        const video = await Video.findById(id);
+        const video = await Video.findById(id)
+        .populate("creator")
+        .populate("comments");// populate는 객체만 가져올수 있다. join과 같은 개념
         res.render("videoDetail", {
             pageTitle: video.title,
             video
@@ -89,7 +94,6 @@ export const videoDetail = async (req, res) => {
 };
 
 export const getEditVideo = async (req, res) => {
-
     const {
         params: {
             id
@@ -97,15 +101,17 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
         const video = await Video.findById(id);
-        res.render("editVideo", {
-            pageTitle: `Edit ${video.title}`,
-            video
-        });
-
+        if(String(video.creator) !== req.user.id){
+            throw Error();
+        }else{
+            res.render("editVideo", {
+                pageTitle: `Edit ${video.title}`,
+                video
+            });
+        }
     } catch (error) {
         res.redirect(routes.home);
     }
-
 }
 
 export const postEditVideo = async (req, res) => {
@@ -120,7 +126,6 @@ export const postEditVideo = async (req, res) => {
     } = req;
 
     try {
-
         await Video.findOneAndUpdate({
             _id: id
         }, {
@@ -133,19 +138,65 @@ export const postEditVideo = async (req, res) => {
     }
 };
 
-
 export const deleteVideo = async (req, res) => {
-
     const {
         params: {
             id
         }
     } = req;
-
     try {
-        await Video.findOneAndRemove({
-            _id: id
-        });
-    } catch (error) {}
+        const video = await Video.findById(id);
+        
+        if(String(video.creator) !== req.user.id){
+            throw Error();
+        }else{
+            await Video.findOneAndRemove({
+                _id: id
+            });
+        }
+    } catch (error) {
+        console.log(error);
+    }
     res.redirect(routes.home);
 }
+
+// Register Video View
+
+export const registerView = async(req, res) => {
+    const {
+        params:{id}
+    }=req;
+    try{
+        const video = await Video.findById(id);
+        video.views+=1;
+        video.save();
+        res.status(200);
+    } catch(error){
+        res.status(400);
+    } finally{
+        res.end();
+    }
+};
+
+// Add Comment
+
+export const postAddComment = async(req,res) => {
+    const{
+        params:{id},
+        body: {comment},
+        user
+    }=req;
+try {
+    const video = await Video.findById(id);
+    const newComment = await Comment.create({
+        text:comment,
+        creator:user.id
+    });  
+    video.comments.push(newComment.id);
+    video.save();
+} catch (error) {
+        res.status(400);
+} finally{
+    res.end();
+}
+};
